refactored useContext; not finished
